tweak error and add test
